Paraná State

A sf object with a polygon for the Paraná state in Brazil. This is a subset of the brazilian map provided by official government agency (IBGE)

Usage

parana

Format

## `parana` A sf with 1 row and 5 columns:

GID0

State code

CODIGOIB1

State's phone code

NOMEUF2

Name of the state

SIGLAUF3

Abbreviation of the state's name

geom

Geometry column of the sf

Source

<https://www.ibge.gov.br/geociencias/cartas-e-mapas/bases-cartograficas-continuas/15759-brasil.html>