Tidyverse methods for caretSDM objects

Set of functions to facilitate the use of caretSDM through tidyverse grammatics.

Usage

select_predictors(x, ...)

# S3 method for class 'sdm_area'
select(x, ...)

# S3 method for class 'input_sdm'
select(x, ...)

# S3 method for class 'sdm_area'
mutate(x, ...)

# S3 method for class 'input_sdm'
mutate(x, ...)

# S3 method for class 'sdm_area'
filter(x, ...)

# S3 method for class 'input_sdm'
filter(x, ...)

# S3 method for class 'occurrences'
filter(x, ...)

filter_species(x, spp = NULL, ...)

Arguments

x

sdm_area object

...

character vector with predictors to be selected.

Examples

# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 25000, crs = 6933)
#> ! Making grid over study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.

# Include predictors:
sa <- add_predictors(sa, bioc) |> dplyr::select(c("bio01", "bio12"))
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
#> Error in dplyr::select(x$grid, ...): Can't select columns that don't exist.
#> ✖ Column `bio01` doesn't exist.