Train SDM models

This function is a wrapper to fit models in caret using caretSDM data.

Usage

train_sdm(occ,
          pred = NULL,
          algo,
          ctrl = NULL,
          variables_selected = NULL,
          parallel = FALSE)

get_tune_length(i)

algorithms_used(i)

get_models(i)

get_validation_metrics(i)

mean_validation_metrics(i)

Arguments

occ

A occurrences or a input_sdm object.

pred

A predictors object. If occ is a input_sdm object, then pred is obtained from it.

algo

A character vector. Algorithms to be used. For a complete list see (https://topepo.github.io/caret/available-models.html). 'maxent', 'mahal.dist' and 'bioclim' are also implemented, but they are only available to run alone, with no other algorithm together.

ctrl

A trainControl object to be used to build models. See trainControl.

variables_selected

A vector of variables to be used as predictors. If NULL, predictors names from pred will be used. Can also be a selection method (e.g. 'vif').

parallel

Should a paralelization method be used (not yet implemented)?

...

Additional arguments to be passed to caret::train function.

i

A models or a input_sdm object.

Value

A models or a input_sdm object.

Details

The object algorithms has a table comparing algorithms available. If the function detects that the necessary packages are not available it will ask for installation. This will happen just in the first time you use the algorithm.

get_tune_length return the length used in grid-search for tunning.

algorithms_used return the names of the algorithms used in the modeling process.

get_models returns a list with trained models (class train) to each species.

get_validation_metrics return a list with a data.frame to each species with complete values for ROC, Sensitivity, Specificity, with their respectives Standard Deviations (SD) and TSS to each of the algorithms and pseudoabsence datasets used.

mean_validation_metrics return a list with a tibble to each species summarizing values for ROC, Sensitivity, Specificity and TSS to each of the algorithms used.

See also

Author

Luíz Fernando Esser (luizesser@gmail.com) https://luizfesser.wordpress.com

Examples

# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 25000, crs = 6933)
#> ! Making grid over study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.

# Include predictors:
sa <- add_predictors(sa, bioc) |> dplyr::select(c("bio01", "bio12"))
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
#> Error in dplyr::select(x$grid, ...): Can't select columns that don't exist.
#> ✖ Column `bio01` doesn't exist.

# Include scenarios:
sa <- add_scenarios(sa)

# Create occurrences:
oc <- occurrences_sdm(occ, crs = 6933) |> join_area(sa)
#> Warning: Some records from `occ` do not fall in `pred`.
#> ℹ 2 elements from `occ` were excluded.
#> ℹ If this seems too much, check how `occ` and `pred` intersect.

# Create input_sdm:
i <- input_sdm(oc, sa)

# Clean coordinates:
i <- data_clean(i)
#> Cell_ids identified, removing duplicated cell_id.
#> Testing country capitals
#> Removed 0 records.
#> Testing country centroids
#> Removed 0 records.
#> Testing duplicates
#> Removed 0 records.
#> Testing equal lat/lon
#> Removed 0 records.
#> Testing biodiversity institutions
#> Removed 1 records.
#> Testing coordinate validity
#> Removed 0 records.
#> Testing sea coordinates
#> Reading layer `ne_110m_land' from data source `/tmp/RtmpBV0kIY/ne_110m_land.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 127 features and 3 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -90 xmax: 180 ymax: 83.64513
#> Geodetic CRS:  WGS 84
#> Removed 0 records.
#> Predictors identified, procceding with grid filter (removing NA and duplicated data).

# VIF calculation:
i <- vif_predictors(i)
#> Warning: the standard deviation is zero
#> Error in xcor[mx[1], mx[2]]: subscript out of bounds

# Pseudoabsence generation:
i <- pseudoabsence(i, method="bioclim", variables_selected = "vif")
#> Error in pseudoabsence(i, method = "bioclim", variables_selected = "vif"): could not find function "pseudoabsence"

# Custom trainControl:
ctrl_sdm <- caret::trainControl(method = "repeatedcv", number = 4, repeats = 10, classProbs = TRUE,
returnResamp = "all", summaryFunction = summary_sdm, savePredictions = "all")

# Train models:
i <- train_sdm(i, algo = c("nnet", "kknn"), variables_selected = "vif", ctrl=ctrl_sdm)
#> Error in rep("pseudoabsence", nrow(pa)): invalid 'times' argument