Add scenarios to sdm_area

This function includes scenarios in the sdm_area object.

Usage

add_scenarios(sa, scen = NULL, scenarios_names = NULL, pred_as_scen = TRUE,
                     variables_selected = NULL, stationary = NULL)

set_scenarios_names(i, scenarios_names = NULL)

scenarios_names(i)

get_scenarios_data(i)

select_scenarios(i, scenarios_names = NULL)

Arguments

sa

A sdm_area or input_sdm object.

scen

RasterStack, SpatRaster or stars object. If NULL adds predictors as a scenario.

scenarios_names

Character vector with names of scenarios.

pred_as_scen

Logical. If TRUE adds the current predictors as a scenario.

variables_selected

Character vector with variables names in scen to be used as variables. If NULL adds all variables.

stationary

Names of variables from sa that should be used in scenarios as stationary variables.

i

A sdm_area or input_sdm object.

Value

add_scenarios returns the input sdm_area or input_sdm object with a new slot called scenarios with scen data as a list, where each slot of the list holds a scenario and each scenario is a sf object. set_scenarios_names sets new names for scenarios in sdm_area/input_sdm object. scenarios_names returns scenarios' names. get_scenarios_data retrieves scenarios data as a list of sf objects. select_scenarios selects scenarios from sdm_area/input_sdm object.

Details

The function add_scenarios adds scenarios to the sdm_area or input_sdm object. If scen has variables that are not present as predictors the function will use only variables present in both objects. stationary variables are those that don't change through the scenarios. It is useful for hidrological variables in fish habitat modeling, for example (see examples below). When adding multiple scenarios in multiple runs, the function will always add a new "current" scenario. To avoid that, set pred_as_scen = FALSE.

See also

Author

Luíz Fernando Esser (luizesser@gmail.com) https://luizfesser.wordpress.com

Examples

# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 50000, crs = 6933)
#> ! Making grid over study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.

# Include predictors:
sa <- add_predictors(sa, bioc)
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.

# Include scenarios:
sa <- add_scenarios(sa, scen)
#> Warning: Some variables in `variables_selected` are not present in `scen`.
#> ℹ Using only variables present in `scen`: bio1, bio4, and bio12

# Set scenarios names:
sa <- set_scenarios_names(sa, scenarios_names = c("future_1", "future_2", "future_3", "future_4",
                                                  "current"))
scenarios_names(sa)
#> [1] "future_1" "future_2" "future_3" "future_4" "current" 

# Get scenarios data:
scenarios_grid <- get_scenarios_data(sa)
scenarios_grid
#> $future_1
#> Simple feature collection with 96 features and 4 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -5276744 ymin: -3295037 xmax: -4626744 ymax: -2795037
#> Projected CRS: WGS 84 / NSIDC EASE-Grid 2.0 Global
#> First 10 features:
#>    cell_id     bio1     bio4    bio12                       geometry
#> 1        3 27.63333 265.8333 1133.250 POLYGON ((-5176744 -2795037...
#> 2        4 27.31667 265.1167 1067.033 POLYGON ((-5126744 -2795037...
#> 3        5 27.15000 266.4667 1048.550 POLYGON ((-5076744 -2795037...
#> 4        6 27.30000 266.3125 1087.100 POLYGON ((-5026744 -2795037...
#> 5        7 27.30000 268.7000 1132.117 POLYGON ((-4976744 -2795037...
#> 6        8 26.88333 275.7167 1129.167 POLYGON ((-4926744 -2795037...
#> 7       15 27.14444 283.1667 1232.511 POLYGON ((-5226744 -2845037...
#> 8       16 27.08889 277.6667 1141.978 POLYGON ((-5176744 -2845037...
#> 9       17 26.56667 275.3667 1104.367 POLYGON ((-5126744 -2845037...
#> 10      18 25.76667 274.6556 1110.511 POLYGON ((-5076744 -2845037...
#> 
#> $future_2
#> Simple feature collection with 96 features and 4 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -5276744 ymin: -3295037 xmax: -4626744 ymax: -2795037
#> Projected CRS: WGS 84 / NSIDC EASE-Grid 2.0 Global
#> First 10 features:
#>    cell_id     bio1     bio4     bio12                       geometry
#> 1        3 33.00000 264.5167 1001.7000 POLYGON ((-5176744 -2795037...
#> 2        4 32.68333 265.2000  945.3667 POLYGON ((-5126744 -2795037...
#> 3        5 32.43333 267.0667  929.5000 POLYGON ((-5076744 -2795037...
#> 4        6 32.56250 267.5875  962.1750 POLYGON ((-5026744 -2795037...
#> 5        7 32.55000 268.8833  999.5000 POLYGON ((-4976744 -2795037...
#> 6        8 32.08333 274.0833  993.4000 POLYGON ((-4926744 -2795037...
#> 7       15 32.52222 280.4667 1086.1222 POLYGON ((-5226744 -2845037...
#> 8       16 32.44444 276.9222 1010.5000 POLYGON ((-5176744 -2845037...
#> 9       17 31.90000 274.9444  977.5000 POLYGON ((-5126744 -2845037...
#> 10      18 31.05556 273.8333  981.9000 POLYGON ((-5076744 -2845037...
#> 
#> $future_3
#> Simple feature collection with 96 features and 4 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -5276744 ymin: -3295037 xmax: -4626744 ymax: -2795037
#> Projected CRS: WGS 84 / NSIDC EASE-Grid 2.0 Global
#> First 10 features:
#>    cell_id     bio1     bio4    bio12                       geometry
#> 1        3 24.33333 278.4167 1379.083 POLYGON ((-5176744 -2795037...
#> 2        4 24.03333 273.0500 1292.950 POLYGON ((-5126744 -2795037...
#> 3        5 23.83333 271.0000 1265.700 POLYGON ((-5076744 -2795037...
#> 4        6 24.01250 266.5750 1304.925 POLYGON ((-5026744 -2795037...
#> 5        7 24.06667 269.6333 1343.817 POLYGON ((-4976744 -2795037...
#> 6        8 23.66667 276.6000 1330.817 POLYGON ((-4926744 -2795037...
#> 7       15 23.86667 302.1000 1510.533 POLYGON ((-5226744 -2845037...
#> 8       16 23.81111 293.2444 1394.444 POLYGON ((-5176744 -2845037...
#> 9       17 23.32222 287.4333 1344.767 POLYGON ((-5126744 -2845037...
#> 10      18 22.48889 284.6778 1345.289 POLYGON ((-5076744 -2845037...
#> 
#> $future_4
#> Simple feature collection with 96 features and 4 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -5276744 ymin: -3295037 xmax: -4626744 ymax: -2795037
#> Projected CRS: WGS 84 / NSIDC EASE-Grid 2.0 Global
#> First 10 features:
#>    cell_id     bio1     bio4    bio12                       geometry
#> 1        3 25.80000 290.6167 1451.633 POLYGON ((-5176744 -2795037...
#> 2        4 25.48333 283.4000 1368.350 POLYGON ((-5126744 -2795037...
#> 3        5 25.28333 279.6000 1342.400 POLYGON ((-5076744 -2795037...
#> 4        6 25.46250 274.1000 1387.837 POLYGON ((-5026744 -2795037...
#> 5        7 25.48333 275.4500 1426.517 POLYGON ((-4976744 -2795037...
#> 6        8 25.06667 281.6500 1412.767 POLYGON ((-4926744 -2795037...
#> 7       15 25.27778 312.9778 1590.411 POLYGON ((-5226744 -2845037...
#> 8       16 25.21111 305.4111 1476.967 POLYGON ((-5176744 -2845037...
#> 9       17 24.70000 298.3222 1432.489 POLYGON ((-5126744 -2845037...
#> 10      18 23.85556 293.8222 1437.589 POLYGON ((-5076744 -2845037...
#> 
#> $current
#> Simple feature collection with 97 features and 4 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -5276744 ymin: -3295037 xmax: -4626744 ymax: -2795037
#> Projected CRS: WGS 84 / NSIDC EASE-Grid 2.0 Global
#> First 10 features:
#>    cell_id     bio1 bio4    bio12                       geometry
#> 1        3 22.29992 1309 264.1004 POLYGON ((-5176744 -2795037...
#> 2        4 22.00096 1234 260.8984 POLYGON ((-5126744 -2795037...
#> 3        5 21.84850 1204 261.5507 POLYGON ((-5076744 -2795037...
#> 4        6 22.02189 1236 259.6719 POLYGON ((-5026744 -2795037...
#> 5        7 22.11035 1273 261.0204 POLYGON ((-4976744 -2795037...
#> 6        8 21.87467 1287 268.6069 POLYGON ((-4926744 -2795037...
#> 7       15 21.87878 1442 282.6844 POLYGON ((-5226744 -2845037...
#> 8       16 21.82576 1306 276.7638 POLYGON ((-5176744 -2845037...
#> 9       17 21.33781 1267 273.3400 POLYGON ((-5126744 -2845037...
#> 10      18 20.47047 1272 273.4352 POLYGON ((-5076744 -2845037...
#> 

# Select scenarios:
sa <- select_scenarios(sa, scenarios_names = c("future_1", "future_2"))
sa
#>           caretSDM         
#> ...........................
#> Class                     : sdm_area
#> Extent                    : -5276744 -3295037 -4626744 -2795037 (xmin, xmax, ymin, ymax)
#> CRS                       : WGS 84 / NSIDC EASE- 
#> Resolution                : (50000, 50000) (x, y)
#> Number of Predictors      : 3 
#> Predictors Names          : bio1, bio4, bio12 
#> Number of Scenarios      : 2 
#> Scenarios Names          : future_1, future_2 

# Setting stationary variables in scenarios:
sa <- sdm_area(rivs[c(1:500),], cell_size = 100000, crs = 6933, lines_as_sdm_area = TRUE) |>
  add_predictors(bioc) |>
  add_scenarios(scen, stationary = c("LENGTH_KM", "DIST_DN_KM"))
#> ! Making grid over study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
#> Linking to GEOS 3.12.1, GDAL 3.8.4, PROJ 9.4.0; sf_use_s2() is TRUE
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
sa
#>           caretSDM         
#> ...........................
#> Class                     : sdm_area
#> Extent                    : -5243024 -3003134 -4743024 -2803134 (xmin, xmax, ymin, ymax)
#> CRS                       : EPSG:6933 
#> Resolution                : (1e+05, 1e+05) (x, y)
#> Number of Predictors      : 5 
#> Predictors Names          : LENGTH_KM, DIST_DN_KM, bio1, bio4, bio12 
#> Number of Scenarios      : 5 
#> Scenarios Names          : ca_ssp245_2090, ca_ssp585_2090, mi_ssp245_2090, mi_ssp585_2090, current