Add scenarios to sdm_area

This function includes scenarios in the sdm_area object.

Usage

add_scenarios(sa, scen = NULL, scenarios_names = NULL, pred_as_scen = TRUE,
                     variables_selected = NULL, stationary = NULL)

set_scenarios_names(i, scenarios_names = NULL)

scenarios_names(i)

get_scenarios_data(i)

select_scenarios(i, scenarios_names = NULL)

Arguments

sa

A sdm_area or input_sdm object.

scen

RasterStack, SpatRaster or stars object. If NULL adds predictors as a scenario.

scenarios_names

Character vector with names of scenarios.

pred_as_scen

Logical. If TRUE adds the current predictors as a scenario.

variables_selected

Character vector with variables names in scen to be used as variables. If NULL adds all variables.

stationary

Names of variables from sa that should be used in scenarios as stationary variables.

i

A sdm_area or input_sdm object.

Value

add_scenarios returns the input sdm_area or input_sdm object with a new slot called scenarios with scen data as a list, where each slot of the list holds a scenario and each scenario is a sf object. set_scenarios_names sets new names for scenarios in sdm_area/input_sdm object. scenarios_names returns scenarios' names. get_scenarios_data retrieves scenarios data as a list of sf objects. select_scenarios selects scenarios from sdm_area/input_sdm object.

Details

The function add_scenarios adds scenarios to the sdm_area or input_sdm object. If scen has variables that are not present as predictors the function will use only variables present in both objects. stationary variables are those that don't change through the scenarios. It is useful for hidrological variables in fish habitat modeling, for example (see examples below). When adding multiple scenarios in multiple runs, the function will always add a new "current" scenario. To avoid that, set pred_as_scen = FALSE.

See also

Author

Luíz Fernando Esser (luizesser@gmail.com) https://luizfesser.wordpress.com

Examples

# Create sdm_area object:
sa <- sdm_area(parana, cell_size = 100000, crs = 6933)
#> ! Making grid over study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.

# Include predictors:
sa <- add_predictors(sa, bioc)
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.

# Include scenarios:
sa <- add_scenarios(sa, scen[1:2]) |> select_predictors(c("bio1", "bio12"))
#> Warning: Some variables in `variables_selected` are not present in `scen`.
#> ℹ Using only variables present in `scen`: bio1, bio4, and bio12

# Set scenarios names:
sa <- set_scenarios_names(sa, scenarios_names = c("future_1", "future_2",
                                                  "current"))
scenarios_names(sa)
#> [1] "future_1" "future_2" "current" 

# Get scenarios data:
scenarios_grid <- get_scenarios_data(sa)
scenarios_grid
#> $future_1
#> Simple feature collection with 26 features and 3 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -5301744 ymin: -3295037 xmax: -4701744 ymax: -2795037
#> Projected CRS: WGS 84 / NSIDC EASE-Grid 2.0 Global
#> First 10 features:
#>    cell_id     bio1    bio12                       geometry
#> 1        2 27.18000 1133.393 POLYGON ((-5201744 -2795037...
#> 2        3 26.32000 1094.763 POLYGON ((-5101744 -2795037...
#> 3        4 26.48667 1176.187 POLYGON ((-5001744 -2795037...
#> 4        5 26.08667 1094.780 POLYGON ((-4901744 -2795037...
#> 5        8 26.70000 1357.180 POLYGON ((-5301744 -2895037...
#> 6        9 25.96667 1244.293 POLYGON ((-5201744 -2895037...
#> 7       10 24.76333 1242.727 POLYGON ((-5101744 -2895037...
#> 8       11 24.14000 1335.097 POLYGON ((-5001744 -2895037...
#> 9       12 24.22000 1208.630 POLYGON ((-4901744 -2895037...
#> 10      13 24.13143 1073.974 POLYGON ((-4801744 -2895037...
#> 
#> $future_2
#> Simple feature collection with 26 features and 3 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -5301744 ymin: -3295037 xmax: -4701744 ymax: -2795037
#> Projected CRS: WGS 84 / NSIDC EASE-Grid 2.0 Global
#> First 10 features:
#>    cell_id     bio1     bio12                       geometry
#> 1        2 32.54667 1002.3133 POLYGON ((-5201744 -2795037...
#> 2        3 31.60667  968.6433 POLYGON ((-5101744 -2795037...
#> 3        4 31.71333 1037.1100 POLYGON ((-5001744 -2795037...
#> 4        5 31.21000  954.8067 POLYGON ((-4901744 -2795037...
#> 5        8 32.08857 1194.4057 POLYGON ((-5301744 -2895037...
#> 6        9 31.30667 1099.7000 POLYGON ((-5201744 -2895037...
#> 7       10 30.02667 1097.2733 POLYGON ((-5101744 -2895037...
#> 8       11 29.32333 1175.0833 POLYGON ((-5001744 -2895037...
#> 9       12 29.26333 1048.3767 POLYGON ((-4901744 -2895037...
#> 10      13 28.95429  901.8114 POLYGON ((-4801744 -2895037...
#> 
#> $current
#> Simple feature collection with 27 features and 3 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -5301744 ymin: -3295037 xmax: -4701744 ymax: -2795037
#> Projected CRS: WGS 84 / NSIDC EASE-Grid 2.0 Global
#> First 10 features:
#>    cell_id     bio1    bio12                       geometry
#> 1        2 21.98257 270.4756 POLYGON ((-5201744 -2795037...
#> 2        3 21.23959 267.7715 POLYGON ((-5101744 -2795037...
#> 3        4 21.46965 264.9304 POLYGON ((-5001744 -2795037...
#> 4        5 21.07145 275.6752 POLYGON ((-4901744 -2795037...
#> 5        8 21.50790 308.3119 POLYGON ((-5301744 -2895037...
#> 6        9 20.93666 295.2626 POLYGON ((-5201744 -2895037...
#> 7       10 19.79140 287.9215 POLYGON ((-5101744 -2895037...
#> 8       11 19.24563 278.8340 POLYGON ((-5001744 -2895037...
#> 9       12 19.49505 278.0062 POLYGON ((-4901744 -2895037...
#> 10      13 19.62922 273.3334 POLYGON ((-4801744 -2895037...
#> 

# Select scenarios:
sa <- select_scenarios(sa, scenarios_names = c("future_1"))

# Setting stationary variables in scenarios:
sa <- sdm_area(rivs[c(1:200),], cell_size = 100000, crs = 6933, lines_as_sdm_area = TRUE) |>
  add_predictors(bioc) |>
  add_scenarios(scen, stationary = c("LENGTH_KM", "DIST_DN_KM"))
#> ! Making grid over study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
#> Linking to GEOS 3.12.1, GDAL 3.8.4, PROJ 9.4.0; sf_use_s2() is TRUE
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.
#> ! Making grid over the study area is an expensive task. Please, be patient!
#> ℹ Using GDAL to make the grid and resample the variables.