Caret Algorithms

A data.frame with characteristics of each algorithm available in caretSDM. Each column is a different characteristics. This can be helpful for more experienced modelers select algorithms. See the source for a selection method using this data.

Usage

algorithms

Format

## `algorithms` A data.frame with 230 rows and 60 columns:

X

Algorithms names

Further columns

Algorithms attributes

Source

<https://topepo.github.io/caret/models-clustered-by-tag-similarity.html>